# openapi/lumina-api.yaml
openapi: '3.0.3'
info:
  title: Lumina API
  version: '1.0.0'
  description: |
    Lumina enterprise learning platform API.
    All authenticated endpoints require `Authorization: Bearer <access_token>` (KeyCloak OIDC).
  contact:
    name: Lumina Engineering
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Versioned API root

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: KeyCloak OIDC access token (5-min TTL, auto-refreshed by oidc-spa)

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PaginationMeta:
      type: object
      required: [cursor, hasMore, total]
      properties:
        cursor:
          type: string
          nullable: true
          description: UUID of the last item — pass as `?cursor=` on next request
        hasMore:
          type: boolean
        total:
          type: integer

    # --- Errors ---
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status, requestId, timestamp]
          properties:
            code:
              type: string
              example: COURSE_NOT_FOUND
            message:
              type: string
              example: Course with ID abc123 does not exist
            status:
              type: integer
              example: 404
            requestId:
              type: string
              example: req_7f8a9b2c
            timestamp:
              type: string
              format: date-time

    # --- Users ---
...
    CourseLevel:
      type: string
      enum: [beginner, intermediate, advanced]
    CourseStatus:
      type: string
      enum: [draft, processing, published, archived]
    Course:
      type: object
      required: [id, title, level, status, isMarketplace, createdAt]
      properties:
        id:
          type: string
          format: uuid
        creatorId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        level:
          $ref: '#/components/schemas/CourseLevel'
        category:
          type: string
          nullable: true
        price:
          type: string
          description: Decimal string e.g. "29.99"
        status:
          $ref: '#/components/schemas/CourseStatus'
        isMarketplace:
          type: boolean
        batchCount:
          type: integer
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseCreate:
      type: object
      required: [title, level]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        level:
          $ref: '#/components/schemas/CourseLevel'
        category:
          type: string
        price:
          type: string
          pattern: '^\d+\.\d{2}$'
          default: '0.00'
        isMarketplace:
          type: boolean
          default: false
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        level:
          $ref: '#/components/schemas/CourseLevel'
        category:
          type: string
        price:
          type: string
        status:
          $ref: '#/components/schemas/CourseStatus'
        isMarketplace:
          type: boolean

    # --- Chapters ---
...
    UploadResponse:
      type: object
      required: [fileId, originalName]
      properties:
        fileId:
          type: string
          format: uuid
        originalName:
          type: string
        sizeBytes:
          type: integer

    # --- Marketplace ---
...
    AdminStats:
      type: object
      required: [totalUsers, totalCourses, totalRevenue, activeJobs]
      properties:
        totalUsers:
          type: integer
        totalCourses:
          type: integer
        totalRevenue:
          type: string
        activeJobs:
          type: integer
        usersByRole:
          type: object
          properties:
            student:
              type: integer
            teacher:
              type: integer
            admin:
              type: integer

    # --- Settings ---
...
  parameters:
    cursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Cursor for pagination (UUID of last item from previous page)
    limitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    courseIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated but insufficient role/ownership
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationError:
      description: Request body or params failed Zod validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

# ---------------------------------------------------------------------------
# Global security — overridden per-endpoint where Public
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# PATHS
# ---------------------------------------------------------------------------
paths:
...
  /courses:
    get:
      operationId: listMyCourses
      summary: List authenticated user's courses
      tags: [Courses]
      parameters:
        - $ref: '#/components/parameters/cursorParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: User's courses
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createCourse
      summary: Create a draft course record
      tags: [Courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Draft course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /courses/{id}:
    parameters:
      - $ref: '#/components/parameters/courseIdParam'

    get:
      operationId: getCourse
      summary: Get course (access-controlled by ownership/purchase/subscription)
      tags: [Courses]
      responses:
        '200':
          description: Course detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateCourse
      summary: Update course metadata (owner or admin only)
      tags: [Courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Updated course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteCourse
      summary: Delete course (owner or admin)
      tags: [Courses]
      responses:
        '204':
          description: Course deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /upload:
    post:
      operationId: uploadFile
      summary: Upload source file (PDF, MD, TXT, image) for course generation
      tags: [Upload]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Max 10MB — PDF, MD, TXT, or image
      responses:
        '200':
          description: File processed, fileId returned for generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File exceeds 10MB limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ==========================================================================
  # AI & GENERATION
  # ==========================================================================
...
  /admin/stats:
    get:
      operationId: getAdminStats
      summary: Platform-wide statistics (admin only)
      tags: [Admin]
      responses:
        '200':
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...